breath::utils

Function zen

pub fn zen() -> Result<i32, InquireError>

Interact with the developer.

Panics

On bad input

Errors

If there is an issue reading the command’s stderr output.

If there is an issue executing a command.

If there is an issue creating a file.

If there is an issue clearing the terminal screen.

If there is an issue writing to a file.

Dependencies

crossterm for terminal manipulation.

git for version control.

git-commit-template for commit templates.